Added Frontend + Backend User Authentication
Started working on frontend elements but realized that there I hadn't properly taken care of user authentication in the backend. In the backend, authentication functions have been created that generate a token whenever a user logs in. I'm going to update this function to have an access token and refresh token so that users are logged off after a certain amount of inactivity.